+++ emhub/data/content/dc_raw.py
@@ -178,39 +178,3 @@ def register_content(dc):
     def task_history(**kwargs):
         return {'task_events': dc.app.dm.get_task_history(kwargs['task_id'])}
 
-    @dc.content
-    def workers_frames(**kwargs):
-        # Some optional parameters
-        days = int(kwargs.get('days', 0))
-        td = dt.timedelta(days=days)
-
-        sortKey = kwargs.get('sort', 'ts')
-        reverse = int(kwargs.get('reverse', 1))
-
-        dm = dc.app.dm
-        hosts = dm.get_hosts()
-
-        folderGroups = {}
-        # TODO: Get worker that monitor cluster from config
-        for h, host in hosts.items():
-            ws = dm.get_worker_stream(h)
-            for t in ws.get_all_tasks():
-                if t['name'] == 'frames' and t['status'] == 'pending':
-                    event_id, event = dm.get_task_lastevent(t['id'])
-                    entries = json.loads(event['entries'])
-                    usage = json.loads(event['usage'])
-                    folders = []
-                    now = dm.now()
-                    for e in entries:
-                        ddt = dm.dt_from_timestamp(e['ts'])
-                        if days == 0 or now - ddt < td:
-                            e['modified'] = ddt
-                            folders.append(e)
-
-                    folders.sort(key=lambda f: f[sortKey], reverse=bool(reverse))
-                    folderGroups[h] = {'usage': usage, 'entries': folders}
-                    break
-
-        return {
-            'folderGroups': folderGroups
-        }
